Move PolygonTypes.ts into types folder of map_planning.
