+++ frontend/src/features/map_planning/types/PolygonTypes.ts
